raspberrypi/requirements.txt
requests
keyboard
pyyaml